Allow caller phone number to include dashes in the right places, but strip them out before saving.
Previously the "edit caller" form did not allow dashes (or anything
except numbers), while the dummy data included dashes in the phone
numbers.  This led to the weird situation where dummy data could be
displayed, but not saved again.

This change does not loosen the restrictions on the value at all, it
just makes it a bit nicer for the user.